+++ demisto_sdk/commands/validate/file_validator.py
@@ -98,9 +98,11 @@ class FilesValidator:
 
     def __init__(self, is_backward_check=True, prev_ver=None, use_git=False, only_committed_files=False,
                  print_ignored_files=False, skip_conf_json=True, validate_id_set=False, file_path=None,
-                 validate_all=False, is_private_repo=False, skip_pack_rn_validation=False, print_ignored_errors=False,
+                 validate_all=False, is_external_repo=False, skip_pack_rn_validation=False, print_ignored_errors=False,
                  configuration=Configuration()):
         self.validate_all = validate_all
+        self.skip_docker_checks = True if self.validate_all else False
+
         self.branch_name = ''
         self.use_git = use_git
         self.skip_pack_rn_validation = skip_pack_rn_validation
@@ -122,8 +124,8 @@ class FilesValidator:
         self.file_path = file_path
         self.changed_pack_data = set()
 
-        self.is_private_repo = is_private_repo
-        if is_private_repo:
+        self.is_external_repo = is_external_repo
+        if is_external_repo:
             print('Running in a private repository')
             self.skip_conf_json = True  # private repository don't have conf.json file
 
@@ -368,24 +370,24 @@ class FilesValidator:
             elif checked_type(file_path, YML_INTEGRATION_REGEXES) or file_type == 'integration':
                 integration_validator = IntegrationValidator(structure_validator, ignored_errors=ignored_errors_list,
                                                              print_as_warnings=self.print_ignored_errors,
-                                                             branch_name=self.branch_name)
+                                                             skip_docker_check=self.skip_docker_checks)
                 if self.is_backward_check and not integration_validator.is_backward_compatible():
                     self._is_valid = False
 
-                if not integration_validator.is_valid_file():
+                if not integration_validator.is_valid_file(skip_test_conf=self.skip_conf_json):
                     self._is_valid = False
 
             elif file_type == 'betaintegration':
                 integration_validator = IntegrationValidator(structure_validator, ignored_errors=ignored_errors_list,
                                                              print_as_warnings=self.print_ignored_errors,
-                                                             branch_name=self.branch_name)
+                                                             skip_docker_check=self.skip_docker_checks)
                 if not integration_validator.is_valid_beta_integration():
                     self._is_valid = False
 
             elif checked_type(file_path, [PACKS_SCRIPT_NON_SPLIT_YML_REGEX]):
                 script_validator = ScriptValidator(structure_validator, ignored_errors=ignored_errors_list,
                                                    print_as_warnings=self.print_ignored_errors,
-                                                   branch_name=self.branch_name)
+                                                   skip_docker_check=self.skip_docker_checks)
                 if self.is_backward_check and not script_validator.is_backward_compatible():
                     self._is_valid = False
                 if not script_validator.is_valid_file():
@@ -403,7 +405,7 @@ class FilesValidator:
                 structure_validator.file_path = yml_path
                 script_validator = ScriptValidator(structure_validator, ignored_errors=ignored_errors_list,
                                                    print_as_warnings=self.print_ignored_errors,
-                                                   branch_name=self.branch_name)
+                                                   skip_docker_check=self.skip_docker_checks)
                 if self.is_backward_check and not script_validator.is_backward_compatible():
                     self._is_valid = False
 
@@ -478,7 +480,7 @@ class FilesValidator:
                     self._is_valid = False
 
             elif checked_type(file_path, CHECKED_TYPES_REGEXES):
-                pass
+                click.secho(f'Could not find validations for file {file_path}', fg='yellow')
 
             else:
                 error_message, error_code = Errors.file_type_not_supported()
@@ -561,8 +563,8 @@ class FilesValidator:
             elif checked_type(file_path, YML_INTEGRATION_REGEXES) or file_type == 'integration':
                 integration_validator = IntegrationValidator(structure_validator, ignored_errors=ignored_errors_list,
                                                              print_as_warnings=self.print_ignored_errors,
-                                                             branch_name=self.branch_name)
-                if not integration_validator.is_valid_file(validate_rn=False):
+                                                             skip_docker_check=self.skip_docker_checks)
+                if not integration_validator.is_valid_file(validate_rn=False, skip_test_conf=self.skip_conf_json):
                     self._is_valid = False
 
             elif checked_type(file_path, PACKAGE_SCRIPTS_REGEXES) or file_type == 'script':
@@ -575,7 +577,7 @@ class FilesValidator:
                 structure_validator.file_path = yml_path
                 script_validator = ScriptValidator(structure_validator, ignored_errors=ignored_errors_list,
                                                    print_as_warnings=self.print_ignored_errors,
-                                                   branch_name=self.branch_name)
+                                                   skip_docker_check=self.skip_docker_checks)
 
                 if not script_validator.is_valid_file(validate_rn=False):
                     self._is_valid = False
@@ -583,7 +585,7 @@ class FilesValidator:
             elif file_type == 'betaintegration':
                 integration_validator = IntegrationValidator(structure_validator, ignored_errors=ignored_errors_list,
                                                              print_as_warnings=self.print_ignored_errors,
-                                                             branch_name=self.branch_name)
+                                                             skip_docker_check=self.skip_docker_checks)
                 if not integration_validator.is_valid_beta_integration(validate_rn=False):
                     self._is_valid = False
 
@@ -648,7 +650,7 @@ class FilesValidator:
                     self._is_valid = False
 
             elif checked_type(file_path, CHECKED_TYPES_REGEXES):
-                pass
+                click.secho(f'Could not find validations for file {file_path}', fg='yellow')
 
             else:
                 error_message, error_code = Errors.file_type_not_supported()
@@ -723,7 +725,7 @@ class FilesValidator:
         Args:
             packs: A set of pack paths i.e {Packs/<pack-name1>, Packs/<pack-name2>}
         """
-        click.secho("================= Validating pack unique files =================\n", fg="bright_cyan")
+        click.secho("\n================= Validating pack unique files =================\n", fg="bright_cyan")
         for pack in packs:
             print(f'Validating {pack} unique pack files\n')
             pack_error_ignore_list = self.get_error_ignore_list(pack)
@@ -731,7 +733,7 @@ class FilesValidator:
                                                                    print_as_warnings=self.print_ignored_errors)
             pack_errors = pack_unique_files_validator.validate_pack_unique_files()
             if pack_errors:
-                print_error(pack_errors)
+                click.secho(pack_errors, fg="bright_red")
                 self._is_valid = False
 
     def run_all_validations_on_file(self, file_path: str, file_type: str = None) -> None:
@@ -765,8 +767,8 @@ class FilesValidator:
         elif checked_type(file_path, INTEGRATION_REGXES) or file_type == 'integration':
             integration_validator = IntegrationValidator(structure_validator, ignored_errors=ignored_errors_list,
                                                          print_as_warnings=self.print_ignored_errors,
-                                                         branch_name=self.branch_name)
-            if not integration_validator.is_valid_file(validate_rn=False):
+                                                         skip_docker_check=self.skip_docker_checks)
+            if not integration_validator.is_valid_file(validate_rn=False, skip_test_conf=self.skip_conf_json):
                 self._is_valid = False
 
         elif checked_type(file_path, YML_ALL_SCRIPTS_REGEXES) or file_type == 'script':
@@ -774,7 +776,7 @@ class FilesValidator:
             structure_validator.file_path = file_path
             script_validator = ScriptValidator(structure_validator, ignored_errors=ignored_errors_list,
                                                print_as_warnings=self.print_ignored_errors,
-                                               branch_name=self.branch_name)
+                                               skip_docker_check=self.skip_docker_checks)
 
             if not script_validator.is_valid_file(validate_rn=False):
                 self._is_valid = False
@@ -782,7 +784,7 @@ class FilesValidator:
         elif file_type == 'betaintegration':
             integration_validator = IntegrationValidator(structure_validator, ignored_errors=ignored_errors_list,
                                                          print_as_warnings=self.print_ignored_errors,
-                                                         branch_name=self.branch_name)
+                                                         skip_docker_check=self.skip_docker_checks)
             if not integration_validator.is_valid_beta_integration():
                 self._is_valid = False
 
@@ -837,10 +839,6 @@ class FilesValidator:
             if not classifier_validator.is_valid_classifier(validate_rn=False):
                 self._is_valid = False
 
-        # elif 'CHANGELOG' in file_path:
-        #     # don't check for CHANGELOG files
-        #     pass
-
         elif checked_type(file_path, CHECKED_TYPES_REGEXES):
             click.secho(f'Could not find validations for file {file_path}', fg='yellow')
 
@@ -853,7 +851,7 @@ class FilesValidator:
         click.secho('\n================= Validating all files =================', fg="bright_cyan")
 
         if not skip_conf_json:
-            print('Validating conf.json')
+            print('\nValidating conf.json')
             conf_json_validator = ConfJsonValidator()
             if not conf_json_validator.is_valid_conf_json():
                 self._is_valid = False
@@ -877,7 +875,9 @@ class FilesValidator:
 
                     if os.path.isfile(file_path):
                         is_yml_file = file_path.endswith('.yml') and \
-                            dir_name in (constants.INTEGRATIONS_DIR, constants.SCRIPTS_DIR, constants.PLAYBOOKS_DIR)
+                            dir_name in (constants.INTEGRATIONS_DIR,
+                                         constants.SCRIPTS_DIR,
+                                         constants.PLAYBOOKS_DIR) and (not file_path.endswith('_unified.yml'))
 
                         is_json_file = file_path.endswith('.json') and \
                             dir_name not in (constants.INTEGRATIONS_DIR, constants.SCRIPTS_DIR, constants.PLAYBOOKS_DIR)
@@ -896,7 +896,8 @@ class FilesValidator:
                                 is_yml_file = inner_file_path.endswith('.yml') and \
                                     (f'/{constants.INTEGRATIONS_DIR}/' in inner_file_path or
                                      f'/{constants.SCRIPTS_DIR}/' in inner_file_path or
-                                     f'/{constants.PLAYBOOKS_DIR}/' in inner_file_path)
+                                     f'/{constants.PLAYBOOKS_DIR}/' in inner_file_path) and \
+                                    not inner_file_path.endswith('_unified.yml')
 
                                 is_md_file = inner_file_path.endswith('README.md')
 
@@ -923,6 +924,8 @@ class FilesValidator:
 
     def validate_all_files_schema(self):
         """Validate all files in the repo are in the right format."""
+        click.secho('\n================= Validates all of Content repo directories according '
+                    'to their schemas =================\n', fg='bright_cyan')
         # go over packs
         for pack_name in os.listdir(PACKS_DIR):
             pack_path = os.path.join(PACKS_DIR, pack_name)
@@ -945,14 +948,16 @@ class FilesValidator:
 
                     if os.path.isfile(file_path):
                         is_yml_file = file_path.endswith('.yml') and \
-                            dir_name in (constants.INTEGRATIONS_DIR, constants.SCRIPTS_DIR, constants.PLAYBOOKS_DIR)
+                            dir_name in (constants.INTEGRATIONS_DIR,
+                                         constants.SCRIPTS_DIR,
+                                         constants.PLAYBOOKS_DIR) and not file_path.endswith('_unified.yml')
 
                         is_json_file = file_path.endswith('.json') and \
                             dir_name not in (
                             constants.INTEGRATIONS_DIR, constants.SCRIPTS_DIR, constants.PLAYBOOKS_DIR)
 
                         if is_yml_file or is_json_file:
-                            print("Validating {}".format(file_path))
+                            print("Validating scheme for {}".format(file_path))
                             self.is_backward_check = False  # if not using git, no need for BC checks
                             structure_validator = StructureValidator(file_path, ignored_errors=ignore_errors_list,
                                                                      print_as_warnings=self.print_ignored_errors)
@@ -968,10 +973,11 @@ class FilesValidator:
                                 is_yml_file = inner_file_path.endswith('.yml') and \
                                     (f'/{constants.INTEGRATIONS_DIR}/' in inner_file_path or
                                      f'/{constants.SCRIPTS_DIR}/' in inner_file_path or
-                                     f'/{constants.PLAYBOOKS_DIR}/' in inner_file_path)
+                                     f'/{constants.PLAYBOOKS_DIR}/' in inner_file_path) and \
+                                    not inner_file_path.endswith('_unified.yml')
 
                                 if is_yml_file:
-                                    print("Validating {}".format(inner_file_path))
+                                    print("Validating scheme for {}".format(inner_file_path))
                                     self.is_backward_check = False  # if not using git, no need for BC checks
                                     structure_validator = StructureValidator(inner_file_path,
                                                                              ignored_errors=ignore_errors_list,
@@ -986,6 +992,7 @@ class FilesValidator:
             (bool). Whether the structure is valid or not.
         """
         if self.validate_all:
+            self.validate_all_files_schema()
             self.validate_all_files(self.skip_conf_json)
             return self._is_valid
 
@@ -1003,8 +1010,6 @@ class FilesValidator:
                 self.validate_committed_files()
             else:
                 self.validate_against_previous_version(no_error=True)
-                click.secho('\n================= Validates all of Content repo directories according '
-                            'to their schemas =================\n', fg='bright_cyan')
                 self.validate_all_files_schema()
 
         else: