+++ CHANGELOG.md
@@ -1,9 +1,18 @@
 # Changelog
+* Added a validation for invalid playbookID fields in incidents types files in **validate** command.
+* Added a code formatter for python files.
 * Fixed an issue where new and old classifiers where mixed on validate command.
 * Added *feedIncremental* field to the integration schema.
 * Fixed error in the **upload** command where unified YMLs were not uploaded as expected if the given input was a pack.
 * Fixed an issue where the **secrets** command failed due to a space character in the file name.
 * Ignored RN validation for *NonSupported* pack.
+* You can now ignore IF107, SC100, RP102 error codes in the **validate** command.
+* Fixed an issue where the **download** command was crashing when received as input a JS integration or script.
+* Fixed an issue where **validate** command checked docker image for JS integrations and scripts.
+* **validate** command now checks scheme for reports and connections.
+* Fixed an issue where **validate** command checked docker when running on all files.
+* Fixed an issue where **validate** command did not fail when docker image was not on the latest numeric tag.
+* Fixed an issue where beta integrations were not validated correctly in **validate** command.
 
 #### 1.1.1
 * fixed and issue where file types were not recognized correctly in **validate** command.